requirements.txt
numpy==1.19.3
selenium
requests
PyQt5
openpyxl
pillow
pypiwin32
passlib
winregistry
packaging
opencv-python
PyQtWebKit
PyQtWebEngine
webdrivermanager
pynput
mysql.connector